asset-service/pyproject.toml
[project]
name = "asset-service"
version = "0.1.0"
description = "Asset upload and processing service for Gemini Studio"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "firebase-admin>=6.4.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-videointelligence>=2.13.0",
    "google-cloud-speech>=2.24.0",
    "google-auth>=2.27.0",
    "pillow>=10.2.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asset_service"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
